requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==4.4.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
proglog==0.1.10
pydub==0.25.1
pypdfium2==4.30.1
pyttsx3==2.98
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
Wand==0.6.13
